03 - Restaurador de Imagens com IA/requirements.txt
huggingface-hub==0.20.3
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
py-real-esrgan==2.0.0
requests==2.32.5
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
streamlit==1.50.0
